Run new solve in unit tests
