+++ tests/eko/runner/test_legacy.py
@@ -9,6 +9,8 @@ from eko import EKO
 from eko.io.runcards import TheoryCard
 from eko.quantities.heavy_quarks import QuarkMassScheme
 
+from . import check_shapes
+
 
 def test_raw(theory_card, operator_card, tmp_path):
     """we don't check the content here, but only the shape"""
@@ -44,38 +46,13 @@ def test_mass_scheme(theory_card, operator_card, tmp_path):
         check_shapes(eko_, eko_.xgrid, eko_.xgrid, theory_card, operator_card)
 
 
-def check_shapes(o, txs, ixs, theory_card, operators_card):
-    tpids = len(o.bases.targetpids)
-    ipids = len(o.bases.inputpids)
-    op_shape = (tpids, len(txs), ipids, len(ixs))
-
-    # check output = input
-    np.testing.assert_allclose(o.xgrid.raw, operators_card.xgrid.raw)
-    # targetgrid and inputgrid in the opcard are now ignored, we are testing this
-    np.testing.assert_allclose(
-        o.bases.targetgrid.raw,
-        txs.raw,
-    )
-    np.testing.assert_allclose(o.bases.inputgrid.raw, ixs.raw)
-    np.testing.assert_allclose(o.mu20, operators_card.mu20)
-    # check available operators
-    ~o.operators
-    assert len(o.mu2grid) == len(operators_card.mu2grid)
-    assert list(o.mu2grid) == operators_card.mu2grid
-    for _, ops in o.items():
-        assert ops.operator.shape == op_shape
-        assert ops.error.shape == op_shape
-
-
 def test_vfns(theory_ffns, operator_card, tmp_path):
-    # change targetpids
     path = tmp_path / "eko.tar"
     tc: TheoryCard = theory_ffns(3)
     oc = copy.deepcopy(operator_card)
     tc.heavy.matching_ratios.c = 1.0
     tc.heavy.matching_ratios.b = 1.0
     tc.order = (2, 0)
-    oc.debug.skip_non_singlet = False
     r = eko.runner.legacy.Runner(tc, oc, path=path)
     r.compute()
     with EKO.read(path) as eko_:
